Moved FileInspector to PluginCore.Helpers.
