+++ External/Plugins/ProjectManager/Projects/LibraryAsset.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 using System.Collections;
-using System.Diagnostics;
+using PluginCore.Helpers;
 using ProjectManager.Projects.AS2;
 
 namespace ProjectManager.Projects
