+++ PluginCore/PluginCore/Helpers/FileHelper.cs
@@ -318,10 +318,6 @@ namespace PluginCore.Helpers
             else return true;
         }
 
-        public static bool IsHaxeExtension(string extension)
-        {
-            return extension == ".hx" || extension == ".hxp";
-        }
     }
 
     /// <summary>
