+++ PluginCore/PluginCore/Helpers/FileInspector.cs
@@ -1,38 +1,34 @@
-using System;
 using System.Collections;
 using System.IO;
-using System.Text;
-using ProjectManager.Helpers;
-using PluginCore.Helpers;
 
-namespace ProjectManager.Projects
+namespace PluginCore.Helpers
 {
-	public class FileInspector
-	{
+    public class FileInspector
+    {
         public static string[] ExecutableFileTypes = null;
 
-		public static bool IsActionScript(string path, string ext)
-		{
-			return ext == ".as";
-		}
+        public static bool IsActionScript(string path, string ext)
+        {
+            return ext == ".as";
+        }
 
         public static bool IsFLA(string path, string ext)
         {
             return ext == ".fla";
         }
 
-		public static bool IsActionScript(ICollection paths)
-		{
+        public static bool IsActionScript(ICollection paths)
+        {
             foreach (string path in paths)
             {
                 if (!IsActionScript(path, Path.GetExtension(path).ToLower())) return false;
             }
-			return true;
-		}
+            return true;
+        }
 
         public static bool IsHaxeFile(string path, string ext)
         {
-            return FileHelper.IsHaxeExtension(ext);
+            return ext == ".hx" || ext == ".hxp";
         }
 
         public static bool IsMxml(string path, string ext)
@@ -46,12 +42,12 @@ namespace ProjectManager.Projects
         }
 
         public static bool IsImage(string path, string ext)
-		{
+        {
             return ext == ".png" || ext == ".jpg" || ext == ".jpeg" || ext == ".gif";
-		}
+        }
 
         public static bool IsSwf(string path, string ext)
-		{
+        {
             return ext == ".swf";
         }
 
@@ -66,7 +62,7 @@ namespace ProjectManager.Projects
         }
 
         public static bool IsFont(string path, string ext)
-		{
+        {
             return ext == ".ttf" || ext == ".otf";
         }
 
@@ -76,45 +72,45 @@ namespace ProjectManager.Projects
         }
 
         public static bool IsResource(string path, string ext)
-		{
+        {
             return IsImage(path, ext) || IsSwf(path, ext) || IsFont(path, ext) || IsSound(path, ext);
-		}
+        }
 
-		public static bool IsResource(ICollection paths)
-		{
+        public static bool IsResource(ICollection paths)
+        {
             foreach (string path in paths)
             {
                 if (!IsResource(path, Path.GetExtension(path).ToLower())) return false;
             }
-			return true;
-		}
+            return true;
+        }
 
-		public static bool ShouldUseShellExecute(string path)
-		{
+        public static bool ShouldUseShellExecute(string path)
+        {
             string ext = Path.GetExtension(path).ToLower();
             if (ExecutableFileTypes != null)
             foreach (string type in ExecutableFileTypes)
             {
                 if (type == ext) return true;
             }
-			return false;
-		}
+            return false;
+        }
 
-		public static bool IsHtml(string path, string ext)
-		{
+        public static bool IsHtml(string path, string ext)
+        {
             return ext == ".html" || ext == ".htm" || ext == ".mtt"/*haxe templo*/;
-		}
+        }
 
-		public static bool IsXml(string path, string ext)
-		{
-			// allow for mxml, sxml, asml, etc
+        public static bool IsXml(string path, string ext)
+        {
+            // allow for mxml, sxml, asml, etc
             return (ext == ".xml" || (ext.Length == 5 && ext.EndsWith("ml")));
-		}
+        }
 
-		public static bool IsText(string path, string ext)
-		{
+        public static bool IsText(string path, string ext)
+        {
             return ext == ".txt" || Path.GetFileName(path).StartsWith(".");
-		}
+        }
 
         public static bool IsAS2Project(string path, string ext)
         {
@@ -149,13 +145,13 @@ namespace ProjectManager.Projects
 
         public static bool IsProject(string path, string ext)
         {
-            return ProjectCreator.IsKnownProject(ext) || IsAS2Project(path, ext) || IsAS3Project(path, ext) || IsHaxeProject(path, ext);
+            return IsAS2Project(path, ext) || IsAS3Project(path, ext) || IsHaxeProject(path, ext);
         }
-
+        
         public static bool IsCustomProject(string path, string ext)
         {
             return (!IsAS2Project(path, ext) && !IsAS3Project(path, ext) &&
-                !IsHaxeProject(path, ext)) && ProjectCreator.IsKnownProject(ext);
+                !IsHaxeProject(path, ext));
         }
 
         public static bool IsTemplate(string path, string ext)
